// types.ts
export type UserRole = 'Admin' | 'Analyst' | 'Viewer' | 'FinOps' | 'DataEngineer';
export type UserStatus = 'Active' | 'Invited' | 'Suspended';

export interface User {
    id: string;
    name: string;
    email: string;
    role: UserRole;
    roleTitle?: string;
    status: UserStatus;
    dateAdded: string;
    avatarUrl?: string;
    message?: string;
    cost: number;
    tokens: number;
    usageTokens?: number; // Usage of NAVSAN tokens
    organization?: string; // Organization name for user information
}
// ...
export interface Database {
    id: string;
    name: string;
    sizeGB: number;
    cost: number;
    // Added credits to solve missing property error in DatabasesView.tsx and dummyData.ts
    credits: number;
    tableCount: number;
    userCount: number;
    users: { id: string, name: string }[];
}